support 3.x admin user creation
In 3.x, the localhost exemption was changed: instead of allowing all
operations, it only allows a user to be created on the admin database.
This means using mongodb::db to define the database and user doesn't
work.

Fix this with two changes:

1. The mongodb_user provider no longer raises an error if it can't
   retrieve the user list.  Instead, it returns an empty array.
2. server.pp creates a user only, instead of a user and a database.
   The admin db should already exist, so this isn't a problem.

I'm not particularly happy about 1., but I'm not sure what a better
fix for this would be.